+++ gemma-rest/src/main/resources/restapidocs/CHANGELOG.md
@@ -12,7 +12,7 @@
   is not initialized
 - omit `subsetFactorId` if `subsetFactor` is rendered
 - omit `factorValuesUsedByExperimentalFactorId` in `DifferentialExpressionAnalysisValueObject` when not set
-  omit `charId` in `FactorValueValueObject`
+- omit `charId` in `FactorValueValueObject`
 
 ### Update 2.8.0